refactor: :recycle: big simplification refactoring

+++ dayco/actions/marquerCommeLu.mjs
@@ -1,17 +1,17 @@
 /**
- * @typedef {import('../../framework/models/user-choices.mjs').UserAnswers} UserAnswers
+ * @typedef {import('../types.mjs').DaycoUserAnswers} DaycoUserAnswers
  */
 
 import { storage } from '../storage/storage.mjs'
 import { POST } from '../types.mjs'
 
 /**
- *
- * @param {UserAnswers} userAnswers
+ * @param {DaycoUserAnswers} userAnswers
  * @returns {Promise<void>}
  */
 export const marquerCommeLu = async (userAnswers) => {
-  const nouveauté = userAnswers['nouveautesDeAbonnement']
-
-  storage.deleteThing(POST, nouveauté)
+  const nouveauté = userAnswers.get('nouveautesDeAbonnement')
+  if (nouveauté) {
+    storage.deleteThing(POST, nouveauté)
+  }
 }
