+++ dayco/actions/supprimerAbonnement.mjs
@@ -1,5 +1,5 @@
 /**
- * @typedef {import('../../framework/models/user-choices.mjs').UserAnswers} UserAnswers
+ * @typedef {import('../types.mjs').DaycoUserAnswers} DaycoUserAnswers
  */
 
 import { showToast } from '../../shared/toast/toast.component.mjs'
@@ -7,14 +7,15 @@ import { storage } from '../storage/storage.mjs'
 import { FEED } from '../types.mjs'
 
 /**
- *
- * @param {UserAnswers} userAnswers
+ * @param {DaycoUserAnswers} userAnswers
  * @returns {Promise<void>}
  */
 export const supprimerAbonnement = async (userAnswers) => {
-  const abonnement = userAnswers['abonnements']
+  const abonnement = userAnswers.get('abonnements')
 
-  await storage.deleteThing(FEED, abonnement)
+  if (abonnement) {
+    await storage.deleteThing(FEED, abonnement)
 
-  showToast('Abonnement supprimé ✅')
+    showToast('Abonnement supprimé ✅')
+  }
 }
